Fix Types:: class inheritance issue
