+++ lib/dumbo.rb
@@ -2,18 +2,18 @@ require 'dumbo/version'
 require 'dumbo/db'
 require 'dumbo/pg_object'
 require 'dumbo/type'
+require 'dumbo/types/composite_type'
+require 'dumbo/types/enum_type'
+require 'dumbo/types/range_type'
+require 'dumbo/types/base_type'
 require 'dumbo/function'
 require 'dumbo/cast'
-require 'dumbo/base_type'
 require 'dumbo/aggregate'
-require 'dumbo/composite_type'
 require 'dumbo/dependency_resolver'
-require 'dumbo/enum_type'
 require 'dumbo/extension'
 require 'dumbo/extension_migrator'
 require 'dumbo/extension_version'
 require 'dumbo/operator'
-require 'dumbo/range_type'
 require 'dumbo/version'
 require 'dumbo/binding_loader'
