+++ spec/type_spec.rb
@@ -31,7 +31,7 @@ describe Dumbo::Type do
     let(:type) { extension.types.select { |t| t.name == type_name }.first }
 
     it 'should have a sql representation' do
-      expect(type.to_sql).to eq <<-SQL.gsub(/^ {6}/, '')
+      expect(type.to_sql).to eq <<-SQL.gsub(/^ {4}/, '')
       CREATE TYPE elephant_base(
         INPUT=elephant_in,
         OUTPUT=elephant_out,
@@ -54,10 +54,10 @@ describe Dumbo::Type do
     let(:type_name) { 'elephant_composite' }
 
     it 'should have a sql representation' do
-      expect(type.to_sql).to eq <<-SQL.gsub(/^ {6}/, '')
+      expect(type.to_sql).to eq <<-SQL.gsub(/^ {4}/, '')
       CREATE TYPE elephant_composite AS (
         weight integer,
-        name text
+      name text
       );
       SQL
     end
@@ -69,11 +69,11 @@ describe Dumbo::Type do
     let(:type_name) { 'elephant_range' }
 
     it 'should have a sql representation' do
-      expect(type.to_sql).to eq <<-SQL.gsub(/^ {6}/, '')
+      expect(type.to_sql).to eq <<-SQL.gsub(/^ {4}/, '')
       CREATE TYPE elephant_range AS RANGE (
         SUBTYPE=float8,
-        SUBTYPE_OPCLASS=float8_ops,
-        SUBTYPE_DIFF=float8mi
+      SUBTYPE_OPCLASS=float8_ops,
+      SUBTYPE_DIFF=float8mi
       );
       SQL
     end
@@ -85,11 +85,11 @@ describe Dumbo::Type do
     let(:type_name) { 'elephant_enum' }
 
     it 'should have a sql representation' do
-      expect(type.to_sql).to eq <<-SQL.gsub(/^ {6}/, '')
+      expect(type.to_sql).to eq <<-SQL.gsub(/^ {4}/, '')
       CREATE TYPE elephant_enum AS ENUM (
         'infant',
-        'child',
-        'adult'
+      'child',
+      'adult'
       );
       SQL
     end
